+++ transports/quic/src/connection.rs
@@ -18,352 +18,353 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-//! A single QUIC connection.
-//!
-//! The [`Connection`] struct of this module contains, amongst other things, a
-//! [`quinn_proto::Connection`] state machine and an `Arc<Endpoint>`. This struct is responsible
-//! for communication between quinn_proto's connection and its associated endpoint.
-//! All interactions with a QUIC connection should be done through this struct.
-// TODO: docs
+mod connecting;
+mod state;
 
-use crate::endpoint::{self, ToEndpoint};
-use futures::{channel::mpsc, prelude::*};
-use futures_timer::Delay;
-use libp2p_core::PeerId;
+use futures::{ready, AsyncRead, AsyncWrite};
+use libp2p_core::muxing::{StreamMuxer, StreamMuxerEvent};
+use parking_lot::Mutex;
+use state::Event;
 use std::{
-    net::SocketAddr,
-    task::{Context, Poll},
-    time::Instant,
+    collections::HashMap,
+    io::{self, Write},
+    pin::Pin,
+    sync::Arc,
+    task::{Context, Poll, Waker},
 };
 
-/// Underlying structure for both [`crate::QuicMuxer`] and [`crate::Connecting`].
+pub use connecting::Connecting;
+pub use state::Error;
+pub use state::State;
+
+/// A single, opened QUIC connection within the libp2p architecture.
 ///
-/// Contains everything needed to process a connection with a remote.
-/// Tied to a specific endpoint.
+/// From libp2p's perspective, a connection must implement the [`StreamMuxer`] interface.
+/// This struct implements satisfies this interface by wrapping around the connection [`State`] machine
+/// and driving it accordingly in the [`StreamMuxer::poll`] function.
 #[derive(Debug)]
 pub struct Connection {
-    /// Channel to the endpoint this connection belongs to.
-    endpoint_channel: endpoint::Channel,
-    /// Pending message to be sent to the background task that is driving the endpoint.
-    pending_to_endpoint: Option<ToEndpoint>,
-    /// Events that the endpoint will send in destination to our local [`quinn_proto::Connection`].
-    /// Passed at initialization.
-    from_endpoint: mpsc::Receiver<quinn_proto::ConnectionEvent>,
-
-    /// The QUIC state machine for this specific connection.
-    connection: quinn_proto::Connection,
-    /// Identifier for this connection according to the endpoint. Used when sending messages to
-    /// the endpoint.
-    connection_id: quinn_proto::ConnectionHandle,
-    /// `Future` that triggers at the [`Instant`] that `self.connection.poll_timeout()` indicates.
-    next_timeout: Option<(Delay, Instant)>,
+    inner: Arc<Mutex<Inner>>,
 }
 
-/// Error on the connection as a whole.
-#[derive(Debug, Clone, thiserror::Error)]
-pub enum ConnectionError {
-    /// The task driving the endpoint has crashed.
-    #[error("Endpoint driver crashed")]
-    EndpointDriverCrashed,
+/// Mutex-protected fields of [`Connection`].
+#[derive(Debug)]
+struct Inner {
+    /// Inner connection object that yields events.
+    state: State,
+    /// State of all the substreams that the muxer reports as open.
+    substreams: HashMap<quinn_proto::StreamId, SubstreamState>,
+    /// Waker to wake if a new outbound substream is opened.
+    poll_outbound_waker: Option<Waker>,
+    /// Waker to wake if a new inbound substream was happened.
+    poll_inbound_waker: Option<Waker>,
+    /// Waker to wake if the connection should be polled again.
+    poll_connection_waker: Option<Waker>,
+}
 
-    /// Error in the inner state machine.
-    #[error(transparent)]
-    Quinn(#[from] quinn_proto::ConnectionError),
+/// State of a single substream.
+#[derive(Debug, Default, Clone)]
+struct SubstreamState {
+    /// Waker to wake if the substream becomes readable or stopped.
+    read_waker: Option<Waker>,
+    /// Waker to wake if the substream becomes writable or stopped.
+    write_waker: Option<Waker>,
+    /// Waker to wake if the substream becomes closed or stopped.
+    finished_waker: Option<Waker>,
 }
 
 impl Connection {
-    /// Crate-internal function that builds a [`Connection`] from raw components.
-    ///
-    /// This function assumes that there exists a [`EndpointDriver`](super::endpoint::EndpointDriver)
-    /// that will process the messages sent to `EndpointChannel::to_endpoint` and send us messages
-    /// on `from_endpoint`.
-    ///
-    /// `connection_id` is used to identify the local connection in the messages sent to
-    /// `to_endpoint`.
-    ///
-    /// This function assumes that the [`quinn_proto::Connection`] is completely fresh and none of
-    /// its methods has ever been called. Failure to comply might lead to logic errors and panics.
-    pub fn from_quinn_connection(
-        endpoint_channel: endpoint::Channel,
-        connection: quinn_proto::Connection,
-        connection_id: quinn_proto::ConnectionHandle,
-        from_endpoint: mpsc::Receiver<quinn_proto::ConnectionEvent>,
-    ) -> Self {
-        debug_assert!(!connection.is_closed());
+    /// Crate-internal function that builds a [`Connection`] from our connection state machine.
+    pub(crate) fn new(state: State) -> Self {
         Connection {
-            endpoint_channel,
-            pending_to_endpoint: None,
-            connection,
-            next_timeout: None,
-            from_endpoint,
-            connection_id,
-        }
-    }
-
-    /// The address that the local socket is bound to.
-    pub fn local_addr(&self) -> &SocketAddr {
-        self.endpoint_channel.socket_addr()
-    }
-
-    /// Returns the address of the node we're connected to.
-    pub fn remote_addr(&self) -> SocketAddr {
-        self.connection.remote_address()
-    }
-
-    /// Start closing the connection. A [`ConnectionEvent::ConnectionLost`] event will be
-    /// produced in the future.
-    pub fn close(&mut self) {
-        // We send a dummy `0` error code with no message, as the API of StreamMuxer doesn't
-        // support this.
-        self.connection
-            .close(Instant::now(), From::from(0u32), Default::default());
-    }
-
-    /// Whether the connection is closed.
-    /// A [`ConnectionEvent::ConnectionLost`] event is emitted with details when the
-    /// connection becomes closed.
-    pub fn is_closed(&self) -> bool {
-        self.connection.is_closed()
-    }
-
-    /// Whether there is no longer any need to keep the connection around.
-    /// All drained connections have been closed.
-    pub fn is_drained(&self) -> bool {
-        self.connection.is_drained()
-    }
-
-    /// Pops a new substream opened by the remote.
-    ///
-    /// If `None` is returned, then a [`ConnectionEvent::StreamAvailable`] event will later be
-    /// produced when a substream is available.
-    pub fn accept_substream(&mut self) -> Option<quinn_proto::StreamId> {
-        self.connection.streams().accept(quinn_proto::Dir::Bi)
-    }
-
-    /// Pops a new substream opened locally.
-    ///
-    /// The API can be thought as if outgoing substreams were automatically opened by the local
-    /// QUIC connection and were added to a queue for availability.
-    ///
-    /// If `None` is returned, then a [`ConnectionEvent::StreamOpened`] event will later be
-    /// produced when a substream is available.
-    pub fn open_substream(&mut self) -> Option<quinn_proto::StreamId> {
-        self.connection.streams().open(quinn_proto::Dir::Bi)
-    }
-
-    /// Control over the stream for reading.
-    pub fn recv_stream(&mut self, id: quinn_proto::StreamId) -> quinn_proto::RecvStream<'_> {
-        self.connection.recv_stream(id)
-    }
-
-    /// Control over the stream for writing.
-    pub fn send_stream(&mut self, id: quinn_proto::StreamId) -> quinn_proto::SendStream<'_> {
-        self.connection.send_stream(id)
-    }
-
-    /// Number of streams that may have unacknowledged data.
-    pub fn send_stream_count(&mut self) -> usize {
-        self.connection.streams().send_streams()
-    }
-
-    /// Closes the given substream.
-    ///
-    /// `write_substream` must no longer be called. The substream is however still
-    /// readable.
-    ///
-    /// On success, a [`quinn_proto::StreamEvent::Finished`] event will later be produced when the
-    /// substream has been effectively closed. A [`ConnectionEvent::StreamStopped`] event can also
-    /// be emitted.
-    pub fn finish_substream(
-        &mut self,
-        id: quinn_proto::StreamId,
-    ) -> Result<(), quinn_proto::FinishError> {
-        self.connection.send_stream(id).finish()
-    }
-
-    /// Polls the connection for an event that happened on it.
-    pub fn poll_event(&mut self, cx: &mut Context<'_>) -> Poll<ConnectionEvent> {
-        loop {
-            match self.from_endpoint.poll_next_unpin(cx) {
-                Poll::Ready(Some(event)) => {
-                    self.connection.handle_event(event);
-                    continue;
-                }
-                Poll::Ready(None) => {
-                    return Poll::Ready(ConnectionEvent::ConnectionLost(
-                        ConnectionError::EndpointDriverCrashed,
-                    ));
-                }
-                Poll::Pending => {}
-            }
-
-            // Sending the pending event to the endpoint. If the endpoint is too busy, we just
-            // stop the processing here.
-            // We need to be careful to avoid a potential deadlock if both `from_endpoint` and
-            // `to_endpoint` are full. As such, we continue to transfer data from `from_endpoint`
-            // to the `quinn_proto::Connection` (see above).
-            // However we don't deliver substream-related events to the user as long as
-            // `to_endpoint` is full. This should propagate the back-pressure of `to_endpoint`
-            // being full to the user.
-            if let Some(to_endpoint) = self.pending_to_endpoint.take() {
-                match self.endpoint_channel.try_send(to_endpoint, cx) {
-                    Ok(Ok(())) => continue, // The endpoint may send back an event.
-                    Ok(Err(to_endpoint)) => {
-                        self.pending_to_endpoint = Some(to_endpoint);
-                        return Poll::Pending;
-                    }
-                    Err(endpoint::Disconnected {}) => {
-                        return Poll::Ready(ConnectionEvent::ConnectionLost(
-                            ConnectionError::EndpointDriverCrashed,
-                        ));
-                    }
-                }
-            }
-
-            // The maximum amount of segments which can be transmitted in a single Transmit
-            // if a platform supports Generic Send Offload (GSO).
-            // Set to 1 for now since not all platforms support GSO.
-            // TODO: Fix for platforms that support GSO.
-            let max_datagrams = 1;
-            // Poll the connection for packets to send on the UDP socket and try to send them on
-            // `to_endpoint`.
-            if let Some(transmit) = self.connection.poll_transmit(Instant::now(), max_datagrams) {
-                // TODO: ECN bits not handled
-                self.pending_to_endpoint = Some(ToEndpoint::SendUdpPacket(transmit));
-                continue;
-            }
-
-            match self.connection.poll_timeout() {
-                Some(timeout) => match self.next_timeout {
-                    Some((_, when)) if when == timeout => {}
-                    _ => {
-                        let now = Instant::now();
-                        // 0ns if now > when
-                        let duration = timeout.duration_since(now);
-                        let next_timeout = Delay::new(duration);
-                        self.next_timeout = Some((next_timeout, timeout))
-                    }
-                },
-                None => self.next_timeout = None,
-            }
-
-            if let Some((timeout, when)) = self.next_timeout.as_mut() {
-                if timeout.poll_unpin(cx).is_ready() {
-                    self.connection.handle_timeout(*when);
-                    continue;
-                }
-            }
-
-            // The connection also needs to be able to send control messages to the endpoint. This is
-            // handled here, and we try to send them on `to_endpoint` as well.
-            if let Some(event) = self.connection.poll_endpoint_events() {
-                let connection_id = self.connection_id;
-                self.pending_to_endpoint = Some(ToEndpoint::ProcessConnectionEvent {
-                    connection_id,
-                    event,
-                });
-                continue;
-            }
-
-            // The final step consists in handling the events related to the various substreams.
-            if let Some(ev) = self.connection.poll() {
-                let event = self.parse_connection_event(ev);
-                return Poll::Ready(event);
-            }
-
-            return Poll::Pending;
-        }
-    }
-
-    fn parse_connection_event(&self, event: quinn_proto::Event) -> ConnectionEvent {
-        match event {
-            quinn_proto::Event::Connected => {
-                let session = self.connection.crypto_session();
-                let identity = session
-                    .peer_identity()
-                    .expect("connection got identity because it passed TLS handshake; qed");
-                let certificates: Box<Vec<rustls::Certificate>> =
-                    identity.downcast().expect("we rely on rustls feature; qed");
-                let end_entity = certificates
-                    .get(0)
-                    .expect("there should be exactly one certificate; qed");
-                let end_entity_der = end_entity.as_ref();
-                let p2p_cert = crate::tls::certificate::parse_certificate(end_entity_der)
-                    .expect("the certificate was validated during TLS handshake; qed");
-                let peer_id = PeerId::from_public_key(&p2p_cert.extension.public_key);
-                ConnectionEvent::Connected(peer_id)
-            }
-            quinn_proto::Event::Stream(quinn_proto::StreamEvent::Readable { id }) => {
-                ConnectionEvent::StreamReadable(id)
-            }
-            quinn_proto::Event::Stream(quinn_proto::StreamEvent::Writable { id }) => {
-                ConnectionEvent::StreamWritable(id)
-            }
-            quinn_proto::Event::Stream(quinn_proto::StreamEvent::Stopped { id, .. }) => {
-                ConnectionEvent::StreamStopped(id)
-            }
-            quinn_proto::Event::Stream(quinn_proto::StreamEvent::Available {
-                dir: quinn_proto::Dir::Bi,
-            }) => ConnectionEvent::StreamAvailable,
-            quinn_proto::Event::Stream(quinn_proto::StreamEvent::Opened {
-                dir: quinn_proto::Dir::Bi,
-            }) => ConnectionEvent::StreamOpened,
-            quinn_proto::Event::ConnectionLost { reason } => {
-                ConnectionEvent::ConnectionLost(ConnectionError::Quinn(reason))
-            }
-            quinn_proto::Event::Stream(quinn_proto::StreamEvent::Finished { id }) => {
-                ConnectionEvent::StreamFinished(id)
-            }
-            quinn_proto::Event::HandshakeDataReady => ConnectionEvent::HandshakeDataReady,
-            quinn_proto::Event::Stream(quinn_proto::StreamEvent::Opened {
-                dir: quinn_proto::Dir::Uni,
-            })
-            | quinn_proto::Event::Stream(quinn_proto::StreamEvent::Available {
-                dir: quinn_proto::Dir::Uni,
-            })
-            | quinn_proto::Event::DatagramReceived => {
-                unreachable!("We don't use datagrams or unidirectional streams.")
-            }
+            inner: Arc::new(Mutex::new(Inner {
+                state,
+                substreams: Default::default(),
+                poll_outbound_waker: None,
+                poll_inbound_waker: None,
+                poll_connection_waker: None,
+            })),
         }
     }
 }
 
-impl Drop for Connection {
-    fn drop(&mut self) {
-        let to_endpoint = ToEndpoint::ProcessConnectionEvent {
-            connection_id: self.connection_id,
-            event: quinn_proto::EndpointEvent::drained(),
+impl StreamMuxer for Connection {
+    type Substream = Substream;
+    type Error = Error;
+
+    fn poll(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<Result<StreamMuxerEvent, Self::Error>> {
+        let mut inner = self.inner.lock();
+        // Poll the inner [`quinn_proto::Connection`] for events and wake
+        // the wakers of related poll-based methods.
+        while let Poll::Ready(event) = inner.state.poll_event(cx) {
+            match event {
+                Event::Connected(_) | Event::HandshakeDataReady => {
+                    debug_assert!(
+                        false,
+                        "Unexpected event {:?} on established QUIC connection",
+                        event
+                    );
+                }
+                Event::ConnectionLost(err) => return Poll::Ready(Err(err)),
+                Event::StreamOpened => {
+                    if let Some(waker) = inner.poll_outbound_waker.take() {
+                        waker.wake();
+                    }
+                }
+                Event::StreamReadable(substream) => {
+                    if let Some(substream) = inner.substreams.get_mut(&substream) {
+                        if let Some(waker) = substream.read_waker.take() {
+                            waker.wake();
+                        }
+                    }
+                }
+                Event::StreamWritable(substream) => {
+                    if let Some(substream) = inner.substreams.get_mut(&substream) {
+                        if let Some(waker) = substream.write_waker.take() {
+                            waker.wake();
+                        }
+                    }
+                }
+                Event::StreamFinished(substream) | Event::StreamStopped(substream) => {
+                    if let Some(substream) = inner.substreams.get_mut(&substream) {
+                        if let Some(waker) = substream.read_waker.take() {
+                            waker.wake();
+                        }
+                        if let Some(waker) = substream.write_waker.take() {
+                            waker.wake();
+                        }
+                        if let Some(waker) = substream.finished_waker.take() {
+                            waker.wake();
+                        }
+                    }
+                }
+                Event::StreamAvailable => {
+                    if let Some(waker) = inner.poll_inbound_waker.take() {
+                        waker.wake();
+                    }
+                }
+            }
+        }
+        // TODO: If connection migration is enabled (currently disabled) address
+        // change on the connection needs to be handled.
+
+        inner.poll_connection_waker = Some(cx.waker().clone());
+        Poll::Pending
+    }
+
+    fn poll_inbound(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<Result<Self::Substream, Self::Error>> {
+        let mut inner = self.inner.lock();
+
+        let substream_id = match inner.state.accept_substream() {
+            Some(id) => {
+                inner.poll_outbound_waker = None;
+                id
+            }
+            None => {
+                inner.poll_inbound_waker = Some(cx.waker().clone());
+                return Poll::Pending;
+            }
         };
-        self.endpoint_channel.send_on_drop(to_endpoint);
+        inner.substreams.insert(substream_id, Default::default());
+        let substream = Substream::new(substream_id, self.inner.clone());
+
+        Poll::Ready(Ok(substream))
+    }
+
+    fn poll_outbound(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<Result<Self::Substream, Self::Error>> {
+        let mut inner = self.inner.lock();
+        let substream_id = match inner.state.open_substream() {
+            Some(id) => {
+                inner.poll_outbound_waker = None;
+                id
+            }
+            None => {
+                inner.poll_outbound_waker = Some(cx.waker().clone());
+                return Poll::Pending;
+            }
+        };
+        inner.substreams.insert(substream_id, Default::default());
+        let substream = Substream::new(substream_id, self.inner.clone());
+        Poll::Ready(Ok(substream))
+    }
+
+    fn poll_close(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        let Inner {
+            substreams,
+            state: connection,
+            ..
+        } = &mut *self.inner.lock();
+        if connection.is_drained() {
+            return Poll::Ready(Ok(()));
+        }
+
+        if connection.send_stream_count() != 0 {
+            for substream in substreams.keys() {
+                if let Err(e) = connection.finish_substream(*substream) {
+                    log::warn!("substream finish error on muxer close: {}", e);
+                }
+            }
+        }
+        loop {
+            if connection.send_stream_count() == 0 && !connection.is_closed() {
+                connection.close()
+            }
+            if let Event::ConnectionLost(_) = ready!(connection.poll_event(cx)) {
+                return Poll::Ready(Ok(()));
+            }
+        }
     }
 }
 
-/// Event generated by the [`Connection`].
 #[derive(Debug)]
-pub enum ConnectionEvent {
-    /// Now connected to the remote and certificates are available.
-    Connected(PeerId),
+pub struct Substream {
+    id: quinn_proto::StreamId,
+    muxer: Arc<Mutex<Inner>>,
+}
 
-    /// Connection has been closed and can no longer be used.
-    ConnectionLost(ConnectionError),
+impl Substream {
+    fn new(id: quinn_proto::StreamId, muxer: Arc<Mutex<Inner>>) -> Self {
+        Self { id, muxer }
+    }
+}
 
-    /// Generated after [`Connection::accept_substream`] has been called and has returned
-    /// `None`. After this event has been generated, this method is guaranteed to return `Some`.
-    StreamAvailable,
-    /// Generated after [`Connection::open_substream`] has been called and has returned
-    /// `None`. After this event has been generated, this method is guaranteed to return `Some`.
-    StreamOpened,
+impl AsyncRead for Substream {
+    fn poll_read(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        mut buf: &mut [u8],
+    ) -> Poll<io::Result<usize>> {
+        let mut muxer = self.muxer.lock();
 
-    /// Generated after `read_substream` has returned a `Blocked` error.
-    StreamReadable(quinn_proto::StreamId),
-    /// Generated after `write_substream` has returned a `Blocked` error.
-    StreamWritable(quinn_proto::StreamId),
+        let mut stream = muxer.state.recv_stream(self.id);
+        let mut chunks = match stream.read(true) {
+            Ok(chunks) => chunks,
+            Err(quinn_proto::ReadableError::UnknownStream) => {
+                return Poll::Ready(Ok(0));
+            }
+            Err(quinn_proto::ReadableError::IllegalOrderedRead) => {
+                unreachable!(
+                    "Illegal ordered read can only happen if `stream.read(false)` is used."
+                );
+            }
+        };
+        let mut bytes = 0;
+        let mut pending = false;
+        loop {
+            if buf.is_empty() {
+                break;
+            }
+            let chunk = match chunks.next(buf.len()) {
+                Ok(Some(chunk)) => chunk,
+                Ok(None) => break,
+                Err(err @ quinn_proto::ReadError::Reset(_)) => {
+                    return Poll::Ready(Err(io::Error::new(io::ErrorKind::ConnectionReset, err)))
+                }
+                Err(quinn_proto::ReadError::Blocked) => {
+                    pending = true;
+                    break;
+                }
+            };
 
-    /// Generated after [`Connection::finish_substream`] has been called.
-    StreamFinished(quinn_proto::StreamId),
-    /// A substream has been stopped. This concept is similar to the concept of a substream being
-    /// "reset", as in a TCP socket being reset for example.
-    StreamStopped(quinn_proto::StreamId),
+            buf.write_all(&chunk.bytes).expect("enough buffer space");
+            bytes += chunk.bytes.len();
+        }
+        if chunks.finalize().should_transmit() {
+            if let Some(waker) = muxer.poll_connection_waker.take() {
+                waker.wake();
+            }
+        }
+        if pending && bytes == 0 {
+            let substream_state = muxer
+                .substreams
+                .get_mut(&self.id)
+                .expect("known substream; qed");
+            substream_state.read_waker = Some(cx.waker().clone());
+            Poll::Pending
+        } else {
+            Poll::Ready(Ok(bytes))
+        }
+    }
+}
 
-    HandshakeDataReady,
+impl AsyncWrite for Substream {
+    fn poll_write(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &[u8],
+    ) -> Poll<std::io::Result<usize>> {
+        let mut muxer = self.muxer.lock();
+
+        match muxer.state.send_stream(self.id).write(buf) {
+            Ok(bytes) => {
+                if let Some(waker) = muxer.poll_connection_waker.take() {
+                    waker.wake();
+                }
+                Poll::Ready(Ok(bytes))
+            }
+            Err(quinn_proto::WriteError::Blocked) => {
+                let substream = muxer
+                    .substreams
+                    .get_mut(&self.id)
+                    .expect("known substream; qed");
+                substream.write_waker = Some(cx.waker().clone());
+                Poll::Pending
+            }
+            Err(err @ quinn_proto::WriteError::Stopped(_)) => {
+                Poll::Ready(Err(io::Error::new(io::ErrorKind::ConnectionReset, err)))
+            }
+            Err(quinn_proto::WriteError::UnknownStream) => {
+                Poll::Ready(Err(io::ErrorKind::BrokenPipe.into()))
+            }
+        }
+    }
+
+    fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
+        // quinn doesn't support flushing, calling close will flush all substreams.
+        Poll::Ready(Ok(()))
+    }
+
+    fn poll_close(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
+        let mut muxer = self.muxer.lock();
+        match muxer.state.finish_substream(self.id) {
+            Ok(()) => {
+                let substream_state = muxer
+                    .substreams
+                    .get_mut(&self.id)
+                    .expect("Substream is not finished.");
+                substream_state.finished_waker = Some(cx.waker().clone());
+                Poll::Pending
+            }
+            Err(err @ quinn_proto::FinishError::Stopped(_)) => {
+                Poll::Ready(Err(io::Error::new(io::ErrorKind::ConnectionReset, err)))
+            }
+            Err(quinn_proto::FinishError::UnknownStream) => Poll::Ready(Ok(())),
+        }
+    }
+}
+
+impl Drop for Substream {
+    fn drop(&mut self) {
+        let mut muxer = self.muxer.lock();
+        muxer.substreams.remove(&self.id);
+        let _ = muxer.state.recv_stream(self.id).stop(0u32.into());
+        let mut send_stream = muxer.state.send_stream(self.id);
+        match send_stream.finish() {
+            Ok(()) => {}
+            // Already finished or reset, which is fine.
+            Err(quinn_proto::FinishError::UnknownStream) => {}
+            Err(quinn_proto::FinishError::Stopped(reason)) => {
+                let _ = send_stream.reset(reason);
+            }
+        }
+    }
 }
