+++ transports/quic/src/endpoint.rs
@@ -29,12 +29,11 @@
 //! [`EndpointDriver`] for a thorough description.
 
 use crate::{
-    connection::Connection,
+    connection,
     provider::Provider,
     tls,
     transport::{self, SocketFamily},
 };
-
 use bytes::BytesMut;
 use futures::{
     channel::{mpsc, oneshot},
@@ -151,7 +150,7 @@ impl Channel {
     pub fn new_bidirectional<P: Provider>(
         quinn_config: QuinnConfig,
         socket_addr: SocketAddr,
-    ) -> Result<(Self, mpsc::Receiver<Connection>), transport::TransportError> {
+    ) -> Result<(Self, mpsc::Receiver<connection::State>), transport::TransportError> {
         let (new_connections_tx, new_connections_rx) = mpsc::channel(1);
         let endpoint = Self::new::<P>(quinn_config, socket_addr, Some(new_connections_tx))?;
         Ok((endpoint, new_connections_rx))
@@ -172,7 +171,7 @@ impl Channel {
     fn new<P: Provider>(
         quinn_config: QuinnConfig,
         socket_addr: SocketAddr,
-        new_connections: Option<mpsc::Sender<Connection>>,
+        new_connections: Option<mpsc::Sender<connection::State>>,
     ) -> Result<Self, transport::TransportError> {
         // NOT blocking, as per man:bind(2), as we pass an IP address.
         let socket = std::net::UdpSocket::bind(&socket_addr)?;
@@ -263,7 +262,7 @@ pub enum ToEndpoint {
         /// UDP address to connect to.
         addr: SocketAddr,
         /// Channel to return the result of the dialing to.
-        result: oneshot::Sender<Result<Connection, quinn_proto::ConnectError>>,
+        result: oneshot::Sender<Result<connection::State, quinn_proto::ConnectError>>,
     },
     /// Sent by a `quinn_proto` connection when the endpoint needs to process an event generated
     /// by a connection. The event itself is opaque to us. Only `quinn_proto` knows what is in
@@ -378,7 +377,7 @@ pub struct EndpointDriver<P: Provider> {
         HashMap<quinn_proto::ConnectionHandle, mpsc::Sender<quinn_proto::ConnectionEvent>>,
     // Channel to forward new inbound connections to the transport.
     // `None` if server capabilities are disabled, i.e. the endpoint is only used for dialing.
-    new_connection_tx: Option<mpsc::Sender<Connection>>,
+    new_connection_tx: Option<mpsc::Sender<connection::State>>,
     // Whether the transport dropped its handle for this endpoint.
     is_decoupled: bool,
 }
@@ -387,7 +386,7 @@ impl<P: Provider> EndpointDriver<P> {
     fn new(
         endpoint_config: Arc<quinn_proto::EndpointConfig>,
         client_config: quinn_proto::ClientConfig,
-        new_connection_tx: Option<mpsc::Sender<Connection>>,
+        new_connection_tx: Option<mpsc::Sender<connection::State>>,
         server_config: Option<Arc<quinn_proto::ServerConfig>>,
         channel: Channel,
         socket: P,
@@ -424,7 +423,7 @@ impl<P: Provider> EndpointDriver<P> {
 
                 debug_assert_eq!(connection.side(), quinn_proto::Side::Client);
                 let (tx, rx) = mpsc::channel(16);
-                let connection = Connection::from_quinn_connection(
+                let connection = connection::State::from_quinn_connection(
                     self.channel.clone(),
                     connection,
                     connection_id,
@@ -530,8 +529,12 @@ impl<P: Provider> EndpointDriver<P> {
                 };
 
                 let (tx, rx) = mpsc::channel(16);
-                let connection =
-                    Connection::from_quinn_connection(self.channel.clone(), connec, connec_id, rx);
+                let connection = connection::State::from_quinn_connection(
+                    self.channel.clone(),
+                    connec,
+                    connec_id,
+                    rx,
+                );
                 match connection_tx.try_send(connection) {
                     Ok(()) => {
                         self.alive_connections.insert(connec_id, tx);
