+++ transports/quic/src/lib.rs
@@ -59,15 +59,13 @@
 
 mod connection;
 mod endpoint;
-mod muxer;
 mod provider;
 mod tls;
 mod transport;
-mod upgrade;
 
-pub use connection::ConnectionError;
+pub use connection::Connection;
+pub use connection::Substream;
 pub use endpoint::Config;
-pub use muxer::Muxer;
 #[cfg(feature = "async-std")]
 pub use provider::async_std;
 #[cfg(feature = "tokio")]
@@ -75,4 +73,3 @@ pub use provider::tokio;
 pub use provider::Provider;
 pub use quinn_proto::ConnectError as DialError;
 pub use transport::{GenTransport, TransportError};
-pub use upgrade::Connecting;
